Add circle class and change returning value in interface from float to double

+++ TestPracticalTask.Library/IArea.cs
@@ -8,6 +8,6 @@ namespace TestPracticalTask.Library
 {
     public interface IArea
     {
-        float CalculateArea();
+        double CalculateArea();
     }
 }
